chore(refactor): file structure
