+++ src/common/components/ProjectsFilter/MobileDateFilterView.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { Back } from './svgs'
-import DatePicker from './DatePicker'
+import DatePicker from '../DatePicker'
 
 import {
   MobileDateHeader,
@@ -12,7 +12,7 @@ import {
   DoneButton,
   MobileDatePicker,
   MobileDatesMenu,
-} from './Style'
+} from './ProjectsFilter.style'
 
 interface Props {
   startDate: any
