+++ src/common/components/ProjectsFilter/svgs.ts
